+++ libs/component-sample/src/lib/component-sample.module.ts
@@ -14,17 +14,31 @@ import { library } from '@fortawesome/fontawesome-svg-core';
 import { sdsIcons } from '@gsa-sam/components';
 import { IconSampleModule } from './feature/icon-sample/icon-sample.module';
 import { IconSampleComponent } from './feature/icon-sample/icon-sample.component';
+import { SideNavigationSampleModule } from './feature/side-navigation-sample/side-navigation-sample.module';
+import { SideNavigationSampleComponent } from './feature/side-navigation-sample/side-navigation-sample.component';
+import { SearchResultListSampleComponent } from './feature/search-result-list-sample/search-result-list-sample.component';
+import { SearchResultListSampleModule } from './feature/search-result-list-sample/search-result-list-sample.module';
+import { SideMenuFiltersSampleModule } from './feature/side-menu-filters-sample/side-menu-filters-sample.module';
+import { SideMenuFiltersSampleComponent } from './feature/side-menu-filters-sample/side-menu-filters-sample.component';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { FormlyModule } from '@ngx-formly/core';
 
+import { SubheaderSampleComponent } from './feature/subheader-sample/subheader-sample.component';
+import { SubheaderSampleModule } from './feature/subheader-sample/subheader-sample.module';
 const appIcons = {
   // App Specific Icons
- };
+};
 
 export const ROUTES: Routes = [
   { path: 'header', component: HeaderSampleComponent },
   { path: 'footer', component: FooterSampleComponent },
   { path: 'pagination', component: PaginationSampleComponent },
   { path: 'toolbar', component: ToolbarSampleComponent },
-  { path: 'icon', component: IconSampleComponent }
+  { path: 'icon', component: IconSampleComponent },
+  { path: 'sideNav', component: SideNavigationSampleComponent },
+  { path: 'sideFilters', component: SideMenuFiltersSampleComponent },
+  { path: 'searchResultList', component: SearchResultListSampleComponent },
+  { path: 'subheader', component: SubheaderSampleComponent }
 ];
 
 @NgModule({
@@ -32,11 +46,18 @@ export const ROUTES: Routes = [
     CommonModule,
     FontAwesomeModule,
     RouterModule.forChild(ROUTES),
+    FormsModule,
+    FormlyModule,
+    ReactiveFormsModule,
     HeaderSampleModule,
     FooterSampleModule,
     PaginationSampleModule,
     ToolbarSampleModule,
-    IconSampleModule
+    IconSampleModule,
+    SideNavigationSampleModule,
+    SideMenuFiltersSampleModule,
+    SearchResultListSampleModule,
+    SubheaderSampleModule
   ],
   exports: [RouterModule]
 })