+++ libs/component-sample/src/lib/feature/footer-sample/footer-sample.component.ts
@@ -1,4 +1,6 @@
 import { Component, OnInit } from '@angular/core';
+import { FooterModel, NavigationMode } from '@gsa-sam/components';
+import { BehaviorSubject } from 'rxjs';
 
 @Component({
   selector: 'gsa-sam-footer-sample',
@@ -6,42 +8,63 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./footer-sample.component.scss']
 })
 export class FooterSampleComponent implements OnInit {
+  constructor() {}
+  public linkEvent = new BehaviorSubject<object>(null);
 
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-  modelFooter = {
-    footerLogo: {
-      text: 'General Services Administration',
-      imageSourcePath: 'assets/sam-styles/img/logo-gsa.svg',
-      imageAltText: 'GSA Logo'
-    },
+  modelFooter: FooterModel = {
     linkSections: [
       {
-        text: 'About beta.SAM.gov', links: [
-          { text: 'Explore Our Community', route: '/' },
-          { text: 'Release Notes', route: '/' }
+        text: 'About beta.SAM.gov',
+        links: [
+          {
+            text: 'Explore Our Community',
+            route: '/',
+            mode: NavigationMode.INTERNAL
+          },
+          { text: 'Release Notes', route: '/', mode: NavigationMode.INTERNAL }
         ]
       },
       {
-        text: 'Our Partners', links: [
-          { text: 'Aquisition.gov', route: '/' },
-          { text: 'USASpending.gov', route: '/' },
-          { text: 'Grants.gov', route: '/' },
-          { text: 'More Partners', route: '/' },
+        text: 'Our Partners',
+        links: [
+          { text: 'Aquisition.gov', route: '/', mode: NavigationMode.INTERNAL },
+          {
+            text: 'USASpending.gov',
+            route: '/',
+            mode: NavigationMode.INTERNAL
+          },
+          { text: 'Grants.gov', route: '/', mode: NavigationMode.INTERNAL },
+          { text: 'More Partners', route: '/', mode: NavigationMode.INTERNAL }
         ]
       },
       {
-        text: 'Customer Service', links: [
-          { text: 'Learning Center', route: '/' },
-          { text: 'Contact Federal Service Desk', route: '/' },
-          { text: 'Resources', route: '/' },
-          { text: 'Freedom of Information Act', route: '/' },
+        text: 'Customer Service',
+        links: [
+          {
+            text: 'Learning Center',
+            route: '/',
+            mode: NavigationMode.INTERNAL
+          },
+          {
+            text: 'Contact Federal Service Desk',
+            route: '/',
+            mode: NavigationMode.INTERNAL
+          },
+          { text: 'Resources', route: '/', mode: NavigationMode.INTERNAL },
+          {
+            text: 'Freedom of Information Act',
+            route: '/',
+            mode: NavigationMode.INTERNAL
+          }
         ]
       }
     ]
   };
 
+  ngOnInit() {
+    this.linkEvent.subscribe(value => {
+      console.log('Link Event Clicked Change');
+      console.log(value);
+    });
+  }
 }