+++ libs/packages/components/src/lib/icon/index.ts
@@ -6,12 +6,33 @@ import {
   faEdit,
   faComment,
   faTh,
-  faSignOutAlt
+  faSignOutAlt,
+  faCircle,
+  faEllipsisV,
+  faTimes,
+  faSearch
 } from "@fortawesome/free-solid-svg-icons";
 
  // SDS Icons
 import { samBars } from '@gsa-sam/sam-styles/src/js/icons/bars';
 import { samArrowLeft } from '@gsa-sam/sam-styles/src/js/icons/arrowLeft';
+import { samBell } from '@gsa-sam/sam-styles/src/js/icons/bell';
+import { samCalendar } from '@gsa-sam/sam-styles/src/js/icons/calendar';
+import { samTrash } from '@gsa-sam/sam-styles/src/js/icons/trash';
+import { samFilter } from '@gsa-sam/sam-styles/src/js/icons/filter';
+import { samBook } from '@gsa-sam/sam-styles/src/js/icons/book';
+import { samHome } from '@gsa-sam/sam-styles/src/js/icons/home';
+import { samEnvelope } from '@gsa-sam/sam-styles/src/js/icons/envelope';
+import { samNewNote } from '@gsa-sam/sam-styles/src/js/icons/newNote';
+import { samFileChartLine } from '@gsa-sam/sam-styles/src/js/icons/fileChartLine';
+import { samShare } from '@gsa-sam/sam-styles/src/js/icons/share';
+import { samRulerPencil } from '@gsa-sam/sam-styles/src/js/icons/rulerPencil';
+import { samPrint } from '@gsa-sam/sam-styles/src/js/icons/print';
+import { samUser } from '@gsa-sam/sam-styles/src/js/icons/user';
+import { samLink } from '@gsa-sam/sam-styles/src/js/icons/link';
+import { samDownload } from '@gsa-sam/sam-styles/src/js/icons/download';
+import { samInvoice } from '@gsa-sam/sam-styles/src/js/icons/invoice';
+import { samPaperPlane } from '@gsa-sam/sam-styles/src/js/icons/paperPlane';
 
  const _icons = {
   // FontAwesome
@@ -20,9 +41,30 @@ import { samArrowLeft } from '@gsa-sam/sam-styles/src/js/icons/arrowLeft';
   faComment,
   faTh,
   faSignOutAlt,
+  faCircle,
+  faEllipsisV,
+  faTimes,
+  faSearch,
   // SDS Icons
   samBars,
-  samArrowLeft
+  samArrowLeft,
+  samBell,
+  samBook,
+  samCalendar,
+  samDownload,
+  samEnvelope,
+  samFileChartLine,
+  samFilter,
+  samHome,
+  samInvoice,
+  samLink,
+  samNewNote,
+  samPaperPlane,
+  samPrint,
+  samRulerPencil,
+  samShare,
+  samTrash,
+  samUser
 };
 
- export const sdsIcons = _icons; 
+ export const sdsIcons = _icons;