+++ libs/packages/components/src/lib/menu/menu-header.component.ts
@@ -1,0 +1,9 @@
+import { Component, Input } from '@angular/core';
+
+@Component({
+  selector: 'sds-menu-header',
+  templateUrl: 'menu-header.component.html'
+})
+export class SdsMenuHeaderComponent {
+  @Input() hideClose = false;
+}