+++ libs/packages/sam-formly/src/lib/formly-filters/sds-filters.component.ts
@@ -1,0 +1,64 @@
+import {
+  Component, Input, Output,
+  ChangeDetectionStrategy,
+  EventEmitter, Optional, OnInit
+} from '@angular/core';
+import { FormGroup } from '@angular/forms';
+import { FormlyFieldConfig } from '@ngx-formly/core';
+import { Subject } from 'rxjs';
+import { SDSFormlyUpdateComunicationService } from './service/sds-filters-comunication.service';
+
+@Component({
+  selector: 'sds-filters',
+  template: `
+        <sds-accordion-item>
+            <sds-accordion-item-header> {{accordionLabel}} </sds-accordion-item-header>
+            <formly-form [form]="form" (modelChange)="modelChange.next($event)" [fields]="fields" [model]="model"></formly-form>
+        </sds-accordion-item>
+    `,
+  changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class SdsFiltersComponent implements OnInit {
+
+  ngOnInit(): void {
+    this.modelChange.subscribe((change) => {
+      this.filterChange.emit(change);
+      if (this.formlyUpdateComunicationService) {
+        this.formlyUpdateComunicationService.updateFilter(change);
+      }
+    })
+  }
+
+  constructor(@Optional() private formlyUpdateComunicationService: SDSFormlyUpdateComunicationService) { }
+
+  /**
+   * Modeal update
+   */
+  modelChange = new Subject<any>();
+
+  /**
+   * Pass in a Form Group for ReactiveForms Support
+   */
+  @Input() public form: FormGroup;
+
+  /**
+   *  Fields are used to configure the UI components 
+   */
+  @Input() public fields: FormlyFieldConfig[];
+
+  /**
+   *  Model used to display the filter values.
+   */
+  @Input() public model: any;
+
+  /**
+   *  Accordion Label used to display the Accordion header text.
+   */
+  @Input() accordionLabel: string = 'Filters';
+
+  /**
+   *  Emit results when model updated
+   */
+  @Output() filterChange = new EventEmitter<object[]>();
+
+}