+++ libs/packages/sam-formly/src/lib/formly-filters/service/sds-filters-comunication.service.ts
@@ -1,0 +1,23 @@
+import { Subject } from 'rxjs';
+export class SDSFormlyUpdateComunicationService {
+
+    /**
+     * Filter update subject
+     */
+    private updateFilterSource = new Subject<any>();
+
+
+    /**
+     * Observable to subscribe to for get updates of the filter
+     */
+    filterUpdate = this.updateFilterSource.asObservable();
+
+    /**
+     * Update the filter
+     * @param filterData
+     */
+    updateFilter(filterData: any) {
+        this.updateFilterSource.next(filterData);
+    }
+
+}