+++ libs/packages/sam-formly/src/lib/public-api.ts
@@ -1,1 +1,3 @@
-export const test = true;
+export * from './formly/formly.module';
+export * from './formly-filters/sds-filters.module';
+export * from './formly-filters/service/sds-filters-comunication.service';