+++ CHANGELOG.md
@@ -1,5 +1,47 @@
 # Changelog
 
+## v0.5.4 (11/07/2019)
+
+#### closed
+
+- [**closed**] IAE-21714 Filter and Search List Communication [#51](https://github.com/GSA/sam-design-system/pull/51)
+- [**closed**] IAE-18673 Icons library [#27](https://github.com/GSA/sam-design-system/pull/27)
+- [**closed**] Entity list - IAE-19965 [#52](https://github.com/GSA/sam-design-system/pull/52)
+
+---
+
+## v0.5.3 (02/07/2019)
+
+#### closed
+
+- [**closed**] IAE-19418 filter output added [#50](https://github.com/GSA/sam-design-system/pull/50)
+- [**closed**] IAE-19418 sds-filters component added  [#46](https://github.com/GSA/sam-design-system/pull/46)
+- [**closed**] Clean up of search results layout custom page [#49](https://github.com/GSA/sam-design-system/pull/49)
+- [**closed**] Fix subheader title alignment [#47](https://github.com/GSA/sam-design-system/pull/47)
+- [**closed**] IAE-19417 for Formly filters with side menu [#42](https://github.com/GSA/sam-design-system/pull/42)
+- [**closed**] IAE-19416 Feedback  [#45](https://github.com/GSA/sam-design-system/pull/45)
+
+---
+
+## v0.5.1 (18/06/2019)
+
+#### closed
+
+- [**closed**] Add grid container to subheader [#44](https://github.com/GSA/sam-design-system/pull/44)
+
+---
+
+## v0.5.0 (18/06/2019)
+
+#### closed
+
+- [**closed**] Adding subheader layout component [#43](https://github.com/GSA/sam-design-system/pull/43)
+- [**closed**] Entity results list [#38](https://github.com/GSA/sam-design-system/pull/38)
+- [**closed**] Adds subheader component and sample page [#33](https://github.com/GSA/sam-design-system/pull/33)
+- [**closed**] IAE-19650 added sam-formly component [#34](https://github.com/GSA/sam-design-system/pull/34)
+
+---
+
 ## v0.2.1 (08/05/2019)
 
 #### closed